Rehabilitate the old BlazingPizza.Server benchmark app. Make it use the framework built from sources.
