+++ src/Components/benchmarkapps/BlazingPizza.Server/Model/Pizza.cs
@@ -1,42 +1,40 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+using System.Globalization;
 
-namespace BlazingPizza
+namespace BlazingPizza.Server.Model;
+
+/// <summary>
+/// Represents a customized pizza as part of an order
+/// </summary>
+public class Pizza
 {
-    /// <summary>
-    /// Represents a customized pizza as part of an order
-    /// </summary>
-    public class Pizza
+    public const int DefaultSize = 12;
+    public const int MinimumSize = 9;
+    public const int MaximumSize = 17;
+
+    public int Id { get; set; }
+
+    public int OrderId { get; set; }
+
+    public PizzaSpecial Special { get; set; }
+
+    public int SpecialId { get; set; }
+
+    public int Size { get; set; }
+
+    public List<PizzaTopping> Toppings { get; set; }
+
+    public decimal GetBasePrice()
     {
-        public const int DefaultSize = 12;
-        public const int MinimumSize = 9;
-        public const int MaximumSize = 17;
+        return Special.BasePrice * (decimal)Size / DefaultSize;
+    }
 
-        public int Id { get; set; }
+    public decimal GetTotalPrice()
+    {
+        return GetBasePrice() + Toppings.Sum(t => t.Topping.Price);
+    }
 
-        public int OrderId { get; set; }
-
-        public PizzaSpecial Special { get; set; }
-
-        public int SpecialId { get; set; }
-
-        public int Size { get; set; }
-
-        public List<PizzaTopping> Toppings { get; set; }
-
-        public decimal GetBasePrice()
-        {
-            return ((decimal)Size / (decimal)DefaultSize) * Special.BasePrice;
-        }
-
-        public decimal GetTotalPrice()
-        {
-            return GetBasePrice() + Toppings.Sum(t => t.Topping.Price);
-        }
-
-        public string GetFormattedTotalPrice()
-        {
-            return GetTotalPrice().ToString("0.00");
-        }
+    public string GetFormattedTotalPrice()
+    {
+        return GetTotalPrice().ToString("0.00", CultureInfo.CurrentCulture);
     }
 }
