+++ src/Components/benchmarkapps/BlazingPizza.Server/Model/PizzaSpecial.cs
@@ -1,20 +1,21 @@
-﻿namespace BlazingPizza
+using System.Globalization;
+
+namespace BlazingPizza.Server.Model;
+
+/// <summary>
+/// Represents a pre-configured template for a pizza a user can order
+/// </summary>
+public class PizzaSpecial
 {
-    /// <summary>
-    /// Represents a pre-configured template for a pizza a user can order
-    /// </summary>
-    public class PizzaSpecial
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public int BasePrice { get; set; }
+    public int BasePrice { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        public string ImageUrl { get; set; }
+    public string ImageUrl { get; set; }
 
-        public string GetFormattedBasePrice() => BasePrice.ToString("0.00");
-    }
+    public string GetFormattedBasePrice() => BasePrice.ToString("0.00", CultureInfo.CurrentCulture);
 }
