--- src/Components/benchmarkapps/BlazingPizza.Server/Model/UserInfo.cs
@@ -1,9 +1,0 @@
-﻿namespace BlazingPizza
-{
-    public class UserInfo
-    {
-        public bool IsAuthenticated { get; set; }
-
-        public string Name { get; set; }
-    }
-}
