+++ src/Components/benchmarkapps/BlazingPizza.Server/Pages/_Host.cshtml
@@ -1,22 +1,19 @@
 ﻿@page "/"
+@using BlazingPizza.Server
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>BlazorServerSIde</title>
+    <title>BlazingPizza.Server</title>
     <base href="~/" />
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-            asp-fallback-href="css/bootstrap/bootstrap.min.css"
-            asp-fallback-test-class="sr-only" asp-fallback-test-property="position" asp-fallback-test-value="absolute"
-            crossorigin="anonymous"
-            integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"/>
-    <link href="css/site.css" rel="stylesheet" />
+    <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
+    <link rel="stylesheet" href="css/site.css" />
 </head>
 <body>
-    <app>@(await Html.RenderComponentAsync<BlazingPizza.Server.App>(RenderMode.ServerPrerendered))</app>
-
-     <script src="_framework/blazor.server.js"></script>
+    <component type="typeof(App)" render-mode="ServerPrerendered" />
+    <script src="_framework/blazor.server.js"></script>
 </body>
 </html>
