+++ src/Components/benchmarkapps/BlazingPizza.Server/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "BlazingPizza.Server": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:61128;http://localhost:61129"
+    }
+  }
+}
