--- src/Components/benchmarkapps/BlazingPizza.Server/SpecialsService.cs
@@ -1,77 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace BlazingPizza.Server
-{
-    public class SpecialsService
-    {
-        public IList<PizzaSpecial> GetSpecials()
-        {
-            return new PizzaSpecial[]
-            {
-                new PizzaSpecial()
-                {
-                    Name = "Basic Cheese Pizza",
-                    Description = "It's cheesy and delicious. Why wouldn't you want one?",
-                    BasePrice = 9,
-                    ImageUrl = "img/pizzas/cheese.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 2,
-                    Name = "The Baconatorizor",
-                    Description = "It has EVERY kind of bacon",
-                    BasePrice = 12,
-                    ImageUrl = "img/pizzas/bacon.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 3,
-                    Name = "Classic pepperoni",
-                    Description = "It's the pizza you grew up with, but Blazing hot!",
-                    BasePrice = 10,
-                    ImageUrl = "img/pizzas/pepperoni.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 4,
-                    Name = "Buffalo chicken",
-                    Description = "Spicy chicken, hot sauce and bleu cheese, guaranteed to warm you up",
-                    BasePrice = 12,
-                    ImageUrl = "img/pizzas/meaty.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 5,
-                    Name = "Mushroom Lovers",
-                    Description = "It has mushrooms. Isn't that obvious?",
-                    BasePrice = 11,
-                    ImageUrl = "img/pizzas/mushroom.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 6,
-                    Name = "The Brit",
-                    Description = "When in London...",
-                    BasePrice = 10,
-                    ImageUrl = "img/pizzas/brit.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 7,
-                    Name = "Veggie Delight",
-                    Description = "It's like salad, but on a pizza",
-                    BasePrice = 11,
-                    ImageUrl = "img/pizzas/salad.jpg",
-                },
-                new PizzaSpecial()
-                {
-                    Id = 8,
-                    Name = "Margherita",
-                    Description = "Traditional Italian pizza with tomatoes and basil",
-                    BasePrice = 9,
-                    ImageUrl = "img/pizzas/margherita.jpg",
-                },
-            };
-        }
-    }
-}
