+++ src/Components/benchmarkapps/BlazingPizza.Server/_Imports.razor
@@ -5,5 +5,6 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
 @using BlazingPizza.Server
+@using BlazingPizza.Server.Model
 @using BlazingPizza.Server.Shared
 @using System.Security.Claims
